Задача 38. Задайте массив вещественных чисел.  Найдите разницу между максимальным и минимальным элементом массива.

+++ Example053 zadacha38_sem1(5)_homeWork/Program.cs
@@ -2,13 +2,13 @@
 //        Найдите разницу между максимальным и минимальным элементом массива.
 
 int[]arr = new int [9];
-int max = arr[0];
+double max = arr[0];
 for (int i = 0; i < arr.Length; i++)
 {
     arr[i] = new Random().Next(99);
     if(arr[i] > max ) max = arr [i];
 }
-int min = arr[0];
+double min = arr[0];
 for (int i = 0; i < arr.Length; i++)
 {
     if(arr[i] < min ) min = arr [i];
